New Script: Bashable, to use: add to any object with a collider and ridgidbody, set a speed, and watch it fly when the fire hits it. Place the object in the bashable layer as well.
Changes to Enemy and bullet scripts to adjust for the new mechanic. Message now sent to objects called "Hit". Bashable objects and the Fire will send it when they hit something.

Enemies now will die from the Hit function that's called when they recieve the hit message or by bullets which still automatically kill anything that isn't a wall or Bashable object.